3. Platform/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.6.1
pydantic==2.5.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncio==3.4.3
typing-extensions==4.9.0
loguru==0.7.2
tenacity==8.2.3
tiktoken==0.5.2
pyyaml==6.0.1
faiss-cpu==1.7.4
numpy==1.24.3